all: unexport cipher types and fix naming to all caps
This is a two small changes that touch a number of files.  First,
the actual cipher types themselves don't need be exported.  This change
renames them all to be internal types to the package.  Instead, our
New*() functions just return a cipher.Block like the rest of the
crypto/* packages do.

Second, the ARIA and HIGHT modules used 'Aria' and 'Hight' in a few
places (notably their New() routines).  This fixes that to the more
correct ARIA and HIGHT.